Revert "Merge branch 'master' of https://gitlab.univ-nantes.fr/E177646T/googlehashcode2018"
This reverts commit 406d008830517d8a359f658c7e3c4730afb04448, reversing
changes made to ed0e2017fbfd89f0ae3db4dfeef1bc2a694bb898.

+++ GoogleHashCode2018/Building.cpp
@@ -239,13 +239,13 @@ const vector<Coord>& Building::getCases() const
 float Building::getRatio()
 {
 	if(ratio<0)
-		ratio = (float) (nbCells / (rowNum * columnNum));
+		ratio = nbCells / (rowNum * columnNum);
 	return ratio;
 }
 float Building::getDensity()
 {
 	if(density<0)
-		density = (float) (nbCells / extra);
+		density = nbCells/extra;
 	return density;
 }
 unsigned int Building::getNbCells() const
